test chaos

+++ snapshot/src/main/scala/takka/chaos/ChaosMessage.scala
@@ -9,7 +9,7 @@ sealed trait ChaosMessage
  * the PoisonPill message, and has (1-`ratio`) possibility of forwarding 
  * the Propagation message to all of its children. 
  */
-case class Propagation(ratio:Float) extends ChaosMessage
+case class Propagation(ratio:Double) extends ChaosMessage
 
 
 class ChaosException(message:String) extends Exception(message)
